requirements.txt
fastapi
uvicorn
python-dotenv
requests
PyMuPDF
python-docx
langchain
sentence-transformers
faiss-cpu
numpy
google-generativeai
